fix: check post-analyzers for StaticPaths (#8904)
